Trekk ut bruker-modellen til eiga fil saman med andre typar som berre vert brukt av brukarmodellen slik at det blir lettare å finne ut kva typar som er filtertypar og kva som er brukarmodell-typar
