+++ src/components/fargekategori/fargekategori-tabellrad-knapp.tsx
@@ -3,7 +3,8 @@ import {AnyAction} from 'redux';
 import {useDispatch} from 'react-redux';
 import {ThunkDispatch} from 'redux-thunk';
 import {Button} from '@navikt/ds-react';
-import {BrukerModell, Fargekategorinavn} from '../../model-interfaces';
+import {BrukerModell} from '../../typer/bruker-modell';
+import {Fargekategorinavn} from '../../model-interfaces';
 import {FargekategoriPopover} from './fargekategori-popover';
 import {fargekategoriIkonMapper} from './fargekategori-ikon-mapper';
 import {resetFargekategoriStateAction} from '../../ducks/fargekategori';
