+++ src/components/modal/tildel-veileder/tildelveileder.test.ts
@@ -1,4 +1,5 @@
-import {FargekategoriModell, HuskelappModell} from '../../../model-interfaces';
+import {FargekategoriModell} from '../../../model-interfaces';
+import {HuskelappModell} from '../../../typer/bruker-modell';
 import {
     harFargekategoriSomVilBliSlettetFilter,
     harHuskelappSomVilBliSlettetFilter,
