+++ src/components/tabell/brukerfnr.tsx
@@ -1,5 +1,5 @@
 import {BodyShort, CopyButton, Tooltip} from '@navikt/ds-react';
-import {BrukerModell} from '../../model-interfaces';
+import {BrukerModell} from '../../typer/bruker-modell';
 
 interface BrukerFnrProps {
     className?: string;
