+++ src/enhetsportefolje/enhet-brukerpanel.tsx
@@ -3,7 +3,8 @@ import {useDispatch} from 'react-redux';
 import classNames from 'classnames';
 import {Checkbox} from '@navikt/ds-react';
 import {Etiketter} from '../components/tabell/etiketter';
-import {BrukerModell, VeilederModell} from '../model-interfaces';
+import {VeilederModell} from '../model-interfaces';
+import {BrukerModell} from '../typer/bruker-modell';
 import {FiltervalgModell} from '../typer/filtervalg-modell';
 import {Kolonne} from '../ducks/ui/listevisning';
 import {EnhetKolonner} from './enhet-kolonner';
