+++ src/minoversikt/huskelapp/panelvisning/HuskelappForPanel.tsx
@@ -1,6 +1,6 @@
 import {BodyShort, Button, Detail, Heading} from '@navikt/ds-react';
 import {toDatePrettyPrint} from '../../../utils/dato-utils';
-import {BrukerModell, HuskelappModell} from '../../../model-interfaces';
+import {BrukerModell, HuskelappModell} from '../../../typer/bruker-modell';
 import {HuskelappPostitWrapper} from '../huskelapp-wrapper/HuskelappPostitWrapper';
 import {SlettHuskelappKnapp} from '../redigering/SlettHuskelappKnapp';
 import '../huskelapp-wrapper/huskelapp-postitstyling.css';
