+++ src/minoversikt/minoversikt-kolonner.tsx
@@ -18,7 +18,8 @@ import {BrukerFnr} from '../components/tabell/brukerfnr';
 import {UkeKolonne} from '../components/tabell/kolonner/ukekolonne';
 import {avvik14aVedtakAvhengigeFilter, ytelseAapSortering, ytelsevalg} from '../filtrering/filter-konstanter';
 import {DatoKolonne} from '../components/tabell/kolonner/datokolonne';
-import {BarnUnder18Aar, BrukerModell, HovedmalNavn, innsatsgruppeNavn} from '../model-interfaces';
+import {HovedmalNavn, innsatsgruppeNavn} from '../model-interfaces';
+import {BarnUnder18Aar, BrukerModell} from '../typer/bruker-modell';
 import {FiltervalgModell} from '../typer/filtervalg-modell';
 import {Kolonne} from '../ducks/ui/listevisning';
 import {TidKolonne} from '../components/tabell/kolonner/tidkolonne';
